test: add spec with attestation, remove inline tests, create test-utils
