+++ src/program.ts
@@ -74,60 +74,6 @@ function createProgram<S extends Spec>(
   };
 }
 
-// inline test
-
-const isMain = import.meta.filename === process.argv[1];
-if (isMain) {
-  let { Bytes } = await import('o1js');
-
-  const Bytes32 = Bytes(32);
-  const InputData = { age: Field, name: Bytes32 };
-
-  // TODO always include owner pk and verify signature on it
-  const spec = Spec(
-    {
-      signedData: Attestation.signature(InputData),
-      targetAge: Input.public(Field),
-      targetName: Input.constant(Bytes32, Bytes32.fromString('Alice')),
-    },
-    ({ signedData, targetAge, targetName }) => ({
-      assert: Operation.and(
-        Operation.equals(Operation.property(signedData, 'age'), targetAge),
-        Operation.equals(Operation.property(signedData, 'name'), targetName)
-      ),
-      data: Operation.property(signedData, 'age'),
-    })
-  );
-
-  function createAttestation<Data>(type: NestedProvableFor<Data>, data: Data) {
-    let issuer = PrivateKey.randomKeypair();
-    let signature = Signature.create(
-      issuer.privateKey,
-      NestedProvable.get(type).toFields(data)
-    );
-    return { public: issuer.publicKey, private: signature, data };
-  }
-
-  let data = { age: Field(18), name: Bytes32.fromString('Alice') };
-  let signedData = createAttestation(InputData, data);
-
-  let program = createProgram(spec);
-  await program.compile();
-
-  // input types are inferred from spec
-  // TODO leverage `From<>` type to pass in inputs directly as numbers / strings etc
-  let proof = await program.run({ signedData, targetAge: Field(18) });
-
-  Provable.log({
-    publicInput: proof.publicInput,
-    publicOutput: proof.publicOutput,
-  });
-
-  // proof types are inferred from spec
-  proof.publicInput satisfies { signedData: PublicKey; targetAge: Field };
-  proof.publicOutput satisfies Field;
-}
-
 // helper
 
 type GetSpecData<S extends Spec> = S extends Spec<infer Data, any>
